+++ app/processing/delta/calculate-overall-delta.js
@@ -1,0 +1,5 @@
+const calculateOverallDelta = (invoiceLines) => {
+  return invoiceLines.reduce((x, y) => x + y.value, 0)
+}
+
+module.exports = calculateOverallDelta
