+++ app/processing/delta/determine-ledger.js
@@ -1,0 +1,5 @@
+const determineLedger = (overallDelta) => {
+  return overallDelta <= 0 ? 'AR' : 'AP'
+}
+
+module.exports = determineLedger
